### 2016-09-21 #78 @peterjc
Potential conflict of interest: I have contributed to the tool with feedback and some small pull requests, https://github.com/sanger-pathogens/gff3toembl/issues?q=author%3Apeterjc https://github.com/sanger-pathogens/gff3toembl/graphs/contributors - does this exclude me as a potential reviewer?
