### 2016-09-22 #78 @arfon
> Hey I was wondering if I could suggest @peterjc as a reviewer?

Thanks @andrewjpage & @peterjc - I'd rather try and find an alternative reviewer for now given that Peter has contributed to this project (albeit small contributions). Do either of you have alternative reviewers in mind?
