### 2016-09-22 #78 @peterjc
Maybe @tseemann could review https://github.com/sanger-pathogens/gff3toembl for JOSS? He is very familiar with American sister sequence archive run by the NCBI though his work on Prokka https://github.com/tseemann/prokka - even if he does not submit to the European sequence archive himself?
