### 2016-09-22 #80 @whedon
[REVIEW]: GFF3toEMBL: Preparing annotated assemblies for submission to EMBL
**Submitting author:** @andrewjpage (<a href="http://orcid.org/0000-0001-6919-6062">Andrew Page</a>)
**Repository:** <a href="https://github.com/sanger-pathogens/gff3toembl.git" target ="_blank">https://github.com/sanger-pathogens/gff3toembl.git</a>
**Version:** v1.1.0
**Archive:** <a href="https://dx.doi.org/10.6084/m9.figshare.3988815.v1" target ="_blank">https://dx.doi.org/10.6084/m9.figshare.3988815.v1</a>
