### 2016-09-22 #78 @arfon
OK the review is open in https://github.com/openjournals/joss-reviews/issues/80 now. @andersgs @andrewjpage let's head over there.
