### 2016-09-28 sanger-pathogens/gff3toembl#62 @andersgs
linuxbrew recipe needs updating
Current `linuxbrew` recipe only goes to version `1.0.9`. `Docker` version is `1.1.0`.
