### 2016-09-28 sanger-pathogens/gff3toembl#63 @andersgs
mac brew install does not work.
It seems the `libpng` dependency is out of date:

![brew_install](https://cloud.githubusercontent.com/assets/3836603/18910007/747bfe30-85b9-11e6-83a5-170641c48072.jpg)
