### 2016-09-28 #80 @andersgs
Terrific work. Nice that you provide a `Docker`. It would be nicer if it were smaller, but I understand the reasons for using the base image. I just had difficulty with the brew install on the mac (sanger-pathogens/gff3toembl#63), and noticed that the brew recipe is one version behind (sanger-pathogens/gff3toembl#62).

Anders.
