### 2016-09-29 sanger-pathogens/gff3toembl#63 @andrewjpage
Hi Anders,
Could you run "brew update" and try it again?  The latest version of libpng is 1.6.25 and your recipe is trying to download is 1.6.17.
Andrew
