+++ README.md
@@ -29,6 +29,7 @@ gff3toembl only works with Python 2.7 and there are known issues on other versio
 
 - Install [http://brew.sh/](HomeBrew)
 - brew tap homebrew/science
+- brew update
 - brew install https://raw.githubusercontent.com/andrewjpage/homebrew-science/gff3toembl/gff3toembl.rb
 
 If you are using Linux:
@@ -36,6 +37,7 @@ If you are using Linux:
 - Install [http://linuxbrew.sh/](LinuxBrew)
 - brew tap homebrew/science
 - brew tap tseemann/homebrew-bioinformatics-linux
+- brew update
 - brew install gff3toembl
 
 ## Example usage
