### 2016-09-29 sanger-pathogens/gff3toembl#62 @andrewjpage
Version updated in PR https://github.com/tseemann/homebrew-bioinformatics-linux/pull/6
