### 2016-10-05 6422b4f
Update paper.md
