### 2016-10-05 #80 @andersgs
I am satisfied everything works as described in the `README`.
