### 2016-10-05 #80 @arfon
Thanks for the review @andersgs! @andrewjpage at this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
