### 2016-10-06 #80 @andrewjpage
Thanks for all the work on this.  I've deposited the code in Figshare https://dx.doi.org/10.6084/m9.figshare.3988815.v1
