### 2016-10-06 #80 @arfon
Thanks for the review @andersgs.

@andrewjpage your paper is now accepted into JOSS. Your DOI is http://dx.doi.org/10.21105/joss.00080 🎉 🚀 💥
